Se agrega la funcionalidad de borrado a la api.

+++ src/com/dao/PersonaDAOtoHibernateImpl.java
@@ -84,7 +84,7 @@ public class PersonaDAOtoHibernateImpl implements PersonaDao{
 		}
 	}
 	
-	public void deletePersona (Integer id){
+	public void deletePersona (Long id){
 		EntityManagerFactory factory = Persistence.createEntityManagerFactory("hibernatePersistenceUnit");
 		EntityManager manager = factory.createEntityManager();
